+++ sender.c
@@ -55,14 +55,14 @@ static struct sum_struct *receive_sums(int f)
 	s->sums = (struct sum_buf *)malloc(sizeof(s->sums[0])*s->count);
 	if (!s->sums) out_of_memory("receive_sums");
 
-	for (i=0;i<s->count;i++) {
+	for (i=0; i < (int) s->count;i++) {
 		s->sums[i].sum1 = read_int(f);
 		read_buf(f,s->sums[i].sum2,csum_length);
 
 		s->sums[i].offset = offset;
 		s->sums[i].i = i;
 
-		if (i == s->count-1 && s->remainder != 0) {
+		if (i == (int) s->count-1 && s->remainder != 0) {
 			s->sums[i].len = s->remainder;
 		} else {
 			s->sums[i].len = s->n;
